microservices/usage-monitor-service/requirements.txt
uvicorn
fastapi
pydantic
azure-storage-blob